scripts/requirements.txt
# pip install -r scripts/requirements.txt
httpx>=0.27,<1
python-dotenv>=1.0,<2
stellar-sdk>=13.0,<14.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4
